# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Calendar Booking API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /public/event-types:
    get:
      summary: List public event types
      responses:
        "200":
          description: Public event types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicEventType"
        "500":
          $ref: "#/components/responses/InternalError"
  /public/event-types/{eventTypeId}/slots:
    get:
      summary: List available slots for an event type
      parameters:
        - name: eventTypeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Slot"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
  /public/bookings:
    post:
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
  /owner/event-types:
    post:
      summary: Create an event type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventTypeRequest"
      responses:
        "201":
          description: Event type created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventType"
        "400":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /owner/bookings:
    get:
      summary: List bookings
      responses:
        "200":
          description: All bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  schemas:
    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: ok
    PublicEventType:
      type: object
      required:
        - id
        - name
        - description
        - durationMinutes
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          format: int32
    EventType:
      type: object
      required:
        - id
        - ownerId
        - name
        - description
        - durationMinutes
      properties:
        id:
          type: string
        ownerId:
          type: string
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          format: int32
    Slot:
      type: object
      required:
        - startAt
        - endAt
        - isFree
      properties:
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        isFree:
          type: boolean
    Booking:
      type: object
      required:
        - id
        - eventTypeId
        - slot
        - createdAt
      properties:
        id:
          type: string
        eventTypeId:
          type: string
        slot:
          $ref: "#/components/schemas/Slot"
        createdAt:
          type: string
          format: date-time
    CreateBookingRequest:
      type: object
      required:
        - eventTypeId
        - slotStartAt
      properties:
        eventTypeId:
          type: string
        slotStartAt:
          type: string
          format: date-time
    CreateEventTypeRequest:
      type: object
      required:
        - id
        - name
        - description
        - durationMinutes
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          format: int32
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: validation_error
            message: request validation failed
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: not_found
            message: resource not found
    ConflictError:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: conflict
            message: slot is already occupied
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: internal_error
            message: internal server error